requirements.txt
pandas
apache-airflow-providers-amazon